# cryptostockz-back/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: "CryptostockZ"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:10010"
basePath: "/"
...
schemes:
- "https"
- "http"
paths:
...
  /product:
    post:
      x-swagger-router-controller: "product.controller"
      tags:
      - "product"
      summary: "Add a new product to the store"
      description: ""
      operationId: "createProduct"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Product object that needs to be added to the store"
        required: true
        schema:
          $ref: "#/definitions/Product"
      responses:
        "405":
          description: "Invalid input"
    put:
      x-swagger-router-controller: "product.controller"
      tags:
      - "product"
      summary: "Update an existing product"
      description: ""
      operationId: "updateProduct"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Product object that needs to be added to the store"
        required: true
        schema:
          $ref: "#/definitions/Product"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Product not found"
        "405":
          description: "Validation exception"
  /product/{productId}:
    get:
      x-swagger-router-controller: "product.controller"
      tags:
      - "product"
      summary: "Find product by ID"
      description: "Returns a single product"
      operationId: "getProductById"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "productId"
        in: "path"
        description: "ID of product to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Product"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Product not found"
    put:
      x-swagger-router-controller: "product.controller"
      tags:
      - "product"
      summary: "Updates a product with form data"
      description: ""
      operationId: "updateProductWithForm"
      consumes:
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "productId"
        in: "path"
        description: "ID of product that needs to be updated"
        required: true
        type: "integer"
        format: "int64"
      - name: "address"
        in: "formData"
        description: "Updated address of the product"
        required: false
        type: "string"
      - name: "owner_address"
        in: "formData"
        description: "Updated owner_address of the product"
        required: false
        type: "string"
      responses:
        "405":
          description: "Invalid input"
    delete:
      x-swagger-router-controller: "product.controller"
      tags:
      - "product"
      summary: "Deletes a product"
      description: ""
      operationId: "deleteProduct"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "api_key"
        in: "header"
        required: false
        type: "string"
      - name: "productId"
        in: "path"
        description: "Product id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Product not found"
  /product/search:
    get:
      x-swagger-router-controller: "product.controller"
      tags:
      - "product"
      summary: "Find product by ID"
      description: "Returns a single product"
      operationId: "searchProduct"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "productName"
        in: "query"
        description: "Name of product to return"
        required: false
        type: "string"
      - name: "manufacturerId"
        in: "query"
        description: "Name of manufacturer to return"
        required: false
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Product"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Product not found"
...
  /account/products/wish:
    get:
      x-swagger-router-controller: "account.controller"
      tags:
      - "account"
      summary: "Finds user wishlist Products"
      description: ""
      operationId: "getUserWishProducts"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Product"
        "400":
          description: "Invalid status value"
...
  /base/products:
    get:
      x-swagger-router-controller: "base_product.controller"
      tags:
      - "account"
      summary: "Finds all Base Products"
      description: ""
      operationId: "getBaseProducts"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Product"
        "400":
          description: "Invalid status value"
definitions:
...
  Product:
    type: "object"
    properties:
      address:
        type: "string"
      base_product_id:
        type: "integer"
      unique_identificator:
        type: "string"
      manufact_id:
        type: "string"
      owner_address:
        type: "string"
      level:
        type: "integer"
    xml:
      name: "Product"
...